requirements.txt
pandas
fastapi
uvicorn
requests
python-dotenv
